# web/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Zentto ERP API
  version: 2.0.0
  description: |
    API REST de Zentto ERP - Sistema de Gestion Empresarial

    **Modulos principales:**
    - Auth (JWT, RBAC, Permisos granulares)
    - Ventas (Facturas, Pedidos, Cotizaciones, Notas credito/debito)
    - Compras (Ordenes, Recepcion de mercancia, Devoluciones)
    - CxC / CxP (Cuentas por Cobrar/Pagar, Cobros, Pagos)
    - Inventario Avanzado (Seriales, Lotes, Bins, FIFO/LIFO, Valoracion)
    - Logistica (Transportistas, Notas de entrega, Guias de despacho)
    - Bancos (Conciliacion bancaria, Caja Chica, Movimientos)
    - Contabilidad (Asientos, Plan de cuentas, Activos fijos, Fiscal)
    - Nomina (Procesamiento, Vacaciones, Liquidaciones, Batch)
    - RRHH (Utilidades, Fideicomiso, Caja de Ahorro, Salud Ocupacional)
    - CRM (Pipeline, Leads, Actividades, Call Center, Campanas)
    - Manufactura (BOM, Ordenes de produccion, Centros de trabajo)
    - Flota (Vehiculos, Combustible, Mantenimiento, Viajes)
    - POS (Facturacion tactil, Impresora fiscal, Cierre de caja)
    - Restaurante (Mesas, Cocina, Recetas)
    - E-Commerce (Tienda online, Carrito, Checkout)
    - Auditoria (Trail automatico, Alertas, Reportes)
    - Permisos (RBAC granular, Aprobaciones, Restricciones de precio)
    - Payments (Pasarelas de pago, Transacciones)
    - Config, Settings, Media, Supervision, Fiscal
  contact:
    name: Zentto ERP
    url: https://zentto.net
    email: info@zentto.net
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        issues:
          type: object
          additionalProperties: true
      required:
        - error
...
    CrmSavedView:
      type: object
      description: Vista guardada (filtros + columnas + orden) por usuario y entidad CRM.
      properties:
        ViewId:
          type: integer
          format: int64
        CompanyId:
          type: integer
        UserId:
          type: integer
        Entity:
          type: string
          enum: [LEAD, CONTACT, COMPANY, DEAL, ACTIVITY]
        Name:
          type: string
        FilterJson:
          type: object
          nullable: true
          additionalProperties: true
        ColumnsJson:
          nullable: true
          oneOf:
            - type: array
              items: {}
            - type: object
              additionalProperties: true
        SortJson:
          nullable: true
          oneOf:
            - type: array
              items: {}
            - type: object
              additionalProperties: true
        IsShared:
          type: boolean
        IsDefault:
          type: boolean
        IsOwner:
          type: boolean
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
    CrmSavedViewCreate:
      type: object
      required: [entity, name]
      properties:
        entity:
          type: string
          enum: [LEAD, CONTACT, COMPANY, DEAL, ACTIVITY]
        name:
          type: string
          maxLength: 200
        filterJson:
          type: object
          additionalProperties: true
        columnsJson:
          nullable: true
          oneOf:
            - type: array
              items: {}
            - type: object
              additionalProperties: true
        sortJson:
          nullable: true
          oneOf:
            - type: array
              items: {}
            - type: object
              additionalProperties: true
        isShared:
          type: boolean
          default: false
        isDefault:
          type: boolean
          default: false
...
    CrmSavedViewMutationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        viewId:
          type: integer
          format: int64
          nullable: true
...
  responses:
    Unauthorized:
      description: Token requerido o invalido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /v1/crm/saved-views:
    get:
      tags:
        - CRM
      summary: Lista vistas guardadas del usuario + compartidas del tenant
      description: Retorna las vistas del usuario actual mas las vistas con IsShared=TRUE del mismo tenant.
      x-db-agnostic: true
      parameters:
        - in: query
          name: entity
          required: false
          schema:
            type: string
            enum: [LEAD, CONTACT, COMPANY, DEAL, ACTIVITY]
      responses:
        "200":
          description: Lista de vistas guardadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      $ref: "#/components/schemas/CrmSavedView"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - CRM
      summary: Crea una vista guardada
      x-db-agnostic: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmSavedViewCreate"
      responses:
        "201":
          description: Vista creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmSavedViewMutationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...